src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1"
description = "GRPC for coments"
authors = ["Obet3000 <zs20015717@estudiantes.uv.mx>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
protobuf = "^4.25.1"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
